Load README for long_description, and drop dependency on nose.
Le sigh... since there is no stable release of nose for Py3k, I've dropped
nose as a dependency for running unit tests. All tests now use unittest,
and there is a custom test suite builder, and a hack to automatically
build the test extension used by tests/test_extension_interface.py.

The hack for building the extension is to get around the inflexible extension
support in distutils. There isn't a way to just add an extension that isn't
automatically included in the binary distribution, so it had to be done
manually.