enhance test infrastructure
setuptools isn't needed anymore for running the tests, they can now be
run with the run-tests.py script.

setup.py is now using a custom build_ext command (derived from
gevent's build_ext command) that's similar to building in-place, but
which also rebuilds the .so files when using a different python
version.