arm32: make sure gcc uses a frame pointer
This code assumes that there is a frame pointer. When gcc doesn't use a
frame pointer here and __thumb__ is defined, then the code clobbers r7
without declaring it to the compiler. If the compiler uses r7 for
something else, then this causes arbitrary failures, depending on what
the optimizer did.

To work around this, tell gcc to always use a frame pointer.