Remove custom my_build_ext and stop trying to build extensions at test time.
Building the extensions and getting them in the right place is the job of tox.

This includes a temporary minor regression in that the test extensions are now also packaged in the binary distributions. This will be resolved with #189 and #184.

Removes some (maybe all, didnt check yet) uses of distutils so partly addresses #185.

Fixes #187

Specify an image for appveyor that works for Python 2.7 (the right
visual studio is missing on other images)

Add installation of missing python versions.

Use PYTHON, not PYTHON_ROOT on appveyor. The former is more common in my experience.

Use a newer image for 3.9 on appveyor