In Python 3.8, Py_AddPendingCall is broken. Workaround that.
It can require the GIL, even though it isn't supposed to. This is
fixed in 3.9 and was not an issue in 3.7.

The symptom is "Fatal Python error: Python memory allocator called
without holding the GIL"

See https://github.com/python/cpython/issues/81308

Unfortunately, the only workaround I see requires deep magic and duplicating CPython internals.