Fix py312+ crash test skips to correctly check for assertions
Fix the skips for `test_trace_events_multiple_greenlets_switching*`
to correctly detect assertions being enabled independently of
`--with-pydebug`.  Gentoo builds Python `--without-pydebug` but
`--with-assertions`, in order to benefit from assertion-implied safety
checks without the full cost of PyDEBUG.  For the purpose of detecting
it, check for `-DNDEBUG` in optimization flags, as implied by disabling
assertions.  Leave the existing code as a fallback for when build flags
are unavailable, e.g. in Windows packages.

Signed-off-by: Michał Górny <mgorny@gentoo.org>